Выполнена функциональность по созданию режиссера, до загрузки github

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,23 +1,18 @@
 package ru.yandex.practicum.filmorate.service;
 
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Director;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.model.MPA;
 import ru.yandex.practicum.filmorate.storage.*;
 import ru.yandex.practicum.filmorate.storage.mappers.DirectorDbStorage;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @Slf4j
 @Service
@@ -86,6 +81,12 @@ public class FilmService {
 
     // Удаление из лайка фильма
     public Film deleteLike(Long filmId, Long userId) {
+        if (filmId == null || filmId < 0) {
+            throw new NotFoundException("Такой фильм не существует");
+        }
+        if (userId == null || userId < 0) {
+            throw new NotFoundException("Такой пользователь не существует");
+        }
         filmDbStorage.deleteLike(filmId, userId);
         return filmDbStorage.findById(filmId).get();
     }
@@ -99,9 +100,19 @@ public class FilmService {
         }
     }
 
+    //Получение всех фильмов новый тест
+    public Collection<Film> getAllFilmsTest() {
+        Set<Film> listAllFilms = new HashSet<>();
+        List<Film> filmsId = filmDbStorage.getAllFilms();
+        for (Film filmId : filmsId) {
+            Film film = filmDbStorage.getById(filmId.getId());
+            listAllFilms.add(film);
+        }
+        return listAllFilms.stream().toList().reversed();
+    }
+
     // Получение фильма по ID
     public Film getFilmWithId(Long filmId) {
-
         Film film = filmDbStorage.getById(filmId);
         if (film == null) {
             throw new NotFoundException("Фильм не найден");
@@ -148,29 +159,29 @@ public class FilmService {
     }
 
     // Получение списка всех режиссеров(directors)
-    public Collection<Director> getAllDirectors(){
-       return directorDbStorage.getAllDirectors();
+    public Collection<Director> getAllDirectors() {
+        return directorDbStorage.getAllDirectors();
     }
 
     // Получение режиссера(director) по ID
     public Director getDirector(long id) {
-     Optional<Director> optionalDirector = directorDbStorage.getDirector(id);
-     if (optionalDirector.isEmpty()) {
-         throw new NotFoundException("Режиссер не найден");
-     }
+        Optional<Director> optionalDirector = directorDbStorage.getDirector(id);
+        if (optionalDirector.isEmpty()) {
+            throw new NotFoundException("Режиссер не найден");
+        }
         return optionalDirector.get();
     }
 
     //Создание режиссера(director)
-    public Director createDirector (Director director) {
-        if (director.getName() == null || StringUtils.containsAny(director.getName(), " ")) {
+    public Director createDirector(Director director) {
+        if (director.getName() == null || director.getName().isBlank()) {
             throw new ValidationException("Не корректное имя режиссера");
         }
-      return directorDbStorage.createDirector(director);
+        return directorDbStorage.createDirector(director);
     }
 
     // Обновление режиссера (director)
-    public Director uptadeDirector (Director director) {
+    public Director uptadeDirector(Director director) {
         return directorDbStorage.uptadeDirector(director);
     }
 
@@ -178,5 +189,20 @@ public class FilmService {
     public void deleteDirector(Long id) {
         directorDbStorage.deleteDirector(id);
     }
+
+    // Получение фильма отсортированного по дате или лайкам
+    public Collection<Film> getFilmSortByLikesOrYears(Long directorId, String sortBy) {
+        if (getDirector(directorId) == null) {
+            throw new NotFoundException("Режиссер не найден");
+        }
+        if (sortBy.equals("likes")) {
+            return filmDbStorage.getFilmsSortByLikes(directorId);
+        } else if (sortBy.equals("year")) {
+            return filmDbStorage.getFilmsSortByDate(directorId);
+        } else {
+            throw new ValidationException("Указан неправильный тип сортировки");
+        }
+    }
+
 
 }
